challenge_2 commit
response data from a post request from client->server will now send data back in csv report style.  the problem was addressed and fixed when reading in a file that is utf8 -- every character including commas, spaces, tabs, and line breaks are represented as a key in an object.  Need to use JSON.parse to parse the data if desiring to iterate through the "JSON" object.